Separando las notificaciones que dependen de area a la nueva aplicación
